[SPARK-XXXXX][SQL][TESTS] Move SQLTestUtilsBase content into QueryTestBase
Move all methods from `SQLTestUtilsBase` into `QueryTestBase`, making `SQLTestUtilsBase` a thin alias that extends `QueryTestBase`.

This is the first step toward consolidating the test utility class hierarchy by merging `SQLTestUtils` into `QueryTest`.
