+++ sql/core/src/test/scala/org/apache/spark/sql/test/SQLTestUtils.scala
@@ -18,32 +18,19 @@
 package org.apache.spark.sql.test
 
 import java.io.File
-import java.net.URI
 import java.nio.file.Files
-import java.util.{Locale, UUID}
 
 import scala.concurrent.duration._
-import scala.jdk.CollectionConverters._
-import scala.language.implicitConversions
 import scala.util.control.NonFatal
 
-import org.apache.hadoop.fs.Path
 import org.scalactic.source.Position
-import org.scalatest.{BeforeAndAfterAll, Suite, Tag}
-import org.scalatest.concurrent.Eventually
+import org.scalatest.{Suite, Tag}
 
 import org.apache.spark.SparkFunSuite
-import org.apache.spark.sql.{classic, AnalysisException, DataFrame, Row, SparkSession}
-import org.apache.spark.sql.catalyst.FunctionIdentifier
-import org.apache.spark.sql.catalyst.analysis.NoSuchTableException
-import org.apache.spark.sql.catalyst.catalog.SessionCatalog.DEFAULT_DATABASE
+import org.apache.spark.sql.{QueryTest, QueryTestBase, Row}
 import org.apache.spark.sql.catalyst.plans.PlanTest
-import org.apache.spark.sql.catalyst.plans.PlanTestBase
-import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.util._
-import org.apache.spark.sql.execution.FilterExec
 import org.apache.spark.sql.execution.adaptive.DisableAdaptiveExecution
-import org.apache.spark.sql.execution.datasources.DataSourceUtils
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.util.ArrayImplicits._
 import org.apache.spark.util.UninterruptibleThread
@@ -214,286 +201,9 @@ private[sql] trait SQLTestUtils extends SparkFunSuite with SQLTestUtilsBase with
 
 /**
  * Helper trait that can be extended by all external SQL test suites.
- *
- * This allows subclasses to plugin a custom `SQLContext`.
- * To use implicit methods, import `testImplicits._` instead of through the `SQLContext`.
- *
- * Subclasses should *not* create `SQLContext`s in the test suite constructor, which is
- * prone to leaving multiple overlapping [[org.apache.spark.SparkContext]]s in the same JVM.
+ * Now a thin alias for [[QueryTestBase]].
  */
-private[sql] trait SQLTestUtilsBase
-  extends Eventually
-  with BeforeAndAfterAll
-  with SQLTestData
-  with PlanTestBase { self: Suite =>
-
-  protected def sparkContext = spark.sparkContext
-
-  // Shorthand for running a query using our SparkSession
-  protected lazy val sql: String => DataFrame = spark.sql _
-
-  /**
-   * A helper object for importing SQL implicits.
-   *
-   * Note that the alternative of importing `spark.implicits._` is not possible here.
-   * This is because we create the `SparkSession` immediately before the first test is run,
-   * but the implicits import is needed in the constructor.
-   */
-  protected object testImplicits
-    extends classic.SQLImplicits
-      with classic.ClassicConversions
-      with classic.ColumnConversions {
-    override protected def session: classic.SparkSession =
-      self.spark.asInstanceOf[classic.SparkSession]
-    override protected def converter: classic.ColumnNodeToExpressionConverter =
-      self.spark.asInstanceOf[classic.SparkSession].converter
-  }
-
-  protected override def withSQLConf[T](pairs: (String, String)*)(f: => T): T = {
-    SparkSession.setActiveSession(spark)
-    super.withSQLConf(pairs: _*)(f)
-  }
-
-  /**
-   * Drops functions after calling `f`. A function is represented by (functionName, isTemporary).
-   */
-  protected def withUserDefinedFunction(functions: (String, Boolean)*)(f: => Unit): Unit = {
-    try {
-      f
-    } catch {
-      case cause: Throwable => throw cause
-    } finally {
-      // If the test failed part way, we don't want to mask the failure by failing to remove
-      // temp tables that never got created.
-      functions.foreach { case (functionName, isTemporary) =>
-        val withTemporary = if (isTemporary) "TEMPORARY" else ""
-        spark.sql(s"DROP $withTemporary FUNCTION IF EXISTS $functionName")
-        assert(
-          !spark.sessionState.catalog.functionExists(FunctionIdentifier(functionName)),
-          s"Function $functionName should have been dropped. But, it still exists.")
-      }
-    }
-  }
-
-  /**
-   * Drops temporary view `viewNames` after calling `f`.
-   */
-  protected def withTempView(viewNames: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f) {
-      viewNames.foreach { viewName =>
-        try spark.catalog.dropTempView(viewName) catch {
-          // If the test failed part way, we don't want to mask the failure by failing to remove
-          // temp views that never got created.
-          case _: NoSuchTableException =>
-        }
-      }
-    }
-  }
-
-  /**
-   * Drops global temporary view `viewNames` after calling `f`.
-   */
-  protected def withGlobalTempView(viewNames: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f) {
-      viewNames.foreach { viewName =>
-        try spark.catalog.dropGlobalTempView(viewName) catch {
-          // If the test failed part way, we don't want to mask the failure by failing to remove
-          // global temp views that never got created.
-          case _: NoSuchTableException =>
-        }
-      }
-    }
-  }
-
-  /**
-   * Drops table `tableName` after calling `f`.
-   */
-  protected def withTable(tableNames: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f) {
-      tableNames.foreach { name =>
-        spark.sql(s"DROP TABLE IF EXISTS $name")
-      }
-    }
-  }
-
-  /**
-   * Drops view `viewName` after calling `f`.
-   */
-  protected def withView(viewNames: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f)(
-      viewNames.foreach { name =>
-        spark.sql(s"DROP VIEW IF EXISTS $name")
-      }
-    )
-  }
-
-  /**
-   * Drops cache `cacheName` after calling `f`.
-   */
-  protected def withCache(cacheNames: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f) {
-      cacheNames.foreach { cacheName =>
-        try uncacheTable(cacheName) catch {
-          case _: AnalysisException =>
-        }
-      }
-    }
-  }
-
-  // Blocking uncache table for tests
-  protected def uncacheTable(tableName: String): Unit = {
-    val tableIdent = spark.sessionState.sqlParser.parseTableIdentifier(tableName)
-    val cascade = !spark.sessionState.catalog.isTempView(tableIdent)
-    spark.sharedState.cacheManager.uncacheQuery(
-      spark.table(tableName).asInstanceOf[classic.Dataset[_]],
-      cascade = cascade,
-      blocking = true)
-  }
-
-  /**
-   * Creates a temporary database and switches current database to it before executing `f`.  This
-   * database is dropped after `f` returns.
-   *
-   * Note that this method doesn't switch current database before executing `f`.
-   */
-  protected def withTempDatabase(f: String => Unit): Unit = {
-    val dbName = s"db_${UUID.randomUUID().toString.replace('-', '_')}"
-
-    try {
-      spark.sql(s"CREATE DATABASE $dbName")
-    } catch { case cause: Throwable =>
-      fail("Failed to create temporary database", cause)
-    }
-
-    try f(dbName) finally {
-      if (spark.catalog.currentDatabase == dbName) {
-        spark.sql(s"USE $DEFAULT_DATABASE")
-      }
-      spark.sql(s"DROP DATABASE $dbName CASCADE")
-    }
-  }
-
-  /**
-   * Drops database `dbName` after calling `f`.
-   */
-  protected def withDatabase(dbNames: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f) {
-      dbNames.foreach { name =>
-        spark.sql(s"DROP DATABASE IF EXISTS $name CASCADE")
-      }
-      spark.sql(s"USE $DEFAULT_DATABASE")
-    }
-  }
-
-  /**
-   * Drops namespace `namespace` after calling `f`.
-   *
-   * Note that, if you switch current catalog/namespace in `f`, you should switch it back manually.
-   */
-  protected def withNamespace(namespaces: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f) {
-      namespaces.foreach { name =>
-        spark.sql(s"DROP NAMESPACE IF EXISTS $name CASCADE")
-      }
-    }
-  }
-
-  /**
-   * Restores the current catalog/database after calling `f`.
-   */
-  protected def withCurrentCatalogAndNamespace(f: => Unit): Unit = {
-    val curCatalog = sql("select current_catalog()").head().getString(0)
-    val curDatabase = sql("select current_database()").head().getString(0)
-    Utils.tryWithSafeFinally(f) {
-      spark.sql(s"USE $curCatalog.$curDatabase")
-    }
-  }
-
-  /**
-   * Enables Locale `language` before executing `f`, then switches back to the default locale of JVM
-   * after `f` returns.
-   */
-  protected def withLocale(language: String)(f: => Unit): Unit = {
-    val originalLocale = Locale.getDefault
-    try {
-      // Add Locale setting
-      Locale.setDefault(new Locale(language))
-      f
-    } finally {
-      Locale.setDefault(originalLocale)
-    }
-  }
-
-  /**
-   * Drops temporary variable `variableName` after calling `f`.
-   */
-  protected def withSessionVariable(variableNames: String*)(f: => Unit): Unit = {
-    Utils.tryWithSafeFinally(f) {
-      variableNames.foreach { name =>
-        spark.sql(s"DROP TEMPORARY VARIABLE IF EXISTS $name")
-      }
-    }
-  }
-
-  /**
-   * Activates database `db` before executing `f`, then switches back to `default` database after
-   * `f` returns.
-   */
-  protected def activateDatabase(db: String)(f: => Unit): Unit = {
-    spark.sessionState.catalogManager.setCurrentNamespace(Array(db))
-    Utils.tryWithSafeFinally(f)(
-      spark.sessionState.catalogManager.setCurrentNamespace(Array("default")))
-  }
-
-  /**
-   * Strip Spark-side filtering in order to check if a datasource filters rows correctly.
-   */
-  protected def stripSparkFilter(df: DataFrame): DataFrame = {
-    val schema = df.schema
-    val withoutFilters = df.queryExecution.executedPlan.transform {
-      case FilterExec(_, child) => child
-    }
-
-    spark.asInstanceOf[classic.SparkSession]
-      .internalCreateDataFrame(withoutFilters.execute(), schema)
-  }
-
-  /**
-   * Turn a logical plan into a `DataFrame`. This should be removed once we have an easier
-   * way to construct `DataFrame` directly out of local data without relying on implicits.
-   */
-  protected implicit def logicalPlanToSparkQuery(plan: LogicalPlan): classic.DataFrame = {
-    classic.Dataset.ofRows(spark.asInstanceOf[classic.SparkSession], plan)
-  }
-
-
-  /**
-   * This method is used to make the given path qualified, when a path
-   * does not contain a scheme, this path will not be changed after the default
-   * FileSystem is changed.
-   */
-  def makeQualifiedPath(path: String): URI = {
-    val hadoopPath = new Path(path)
-    val fs = hadoopPath.getFileSystem(spark.sessionState.newHadoopConf())
-    fs.makeQualified(hadoopPath).toUri
-  }
-
-  /**
-   * Returns full path to the given file in the resource folder
-   */
-  protected def testFile(fileName: String): String = {
-    Thread.currentThread().getContextClassLoader.getResource(fileName).toString
-  }
-
-  /**
-   * Returns the size of the local directory except the metadata file and the temporary file.
-   */
-  def getLocalDirSize(file: File): Long = {
-    assert(file.isDirectory)
-    Files.walk(file.toPath).iterator().asScala
-      .filter(p => Files.isRegularFile(p) && DataSourceUtils.isDataFile(p.getFileName.toString))
-      .map(_.toFile.length).sum
-  }
+private[sql] trait SQLTestUtilsBase extends QueryTestBase { self: Suite =>
 }
 
 private[sql] object SQLTestUtils {
